+++ jsapp/js/utils.tests.ts
@@ -1,5 +1,6 @@
 import {
   formatSeconds,
+  formatTimeFromSeconds,
   generateAutoname,
   getLangAsObject,
   getLangString,
@@ -168,6 +169,43 @@ describe('utils', () => {
       const INDEX_FIRST_WORD = 4
       const INDEX_LAST_WORD = 21
       chai.expect(generateAutoname(testString, INDEX_FIRST_WORD, INDEX_LAST_WORD)).to.equal('___a_very_long_na')
+    })
+  })
+
+  describe('formatTimeFromSeconds', () => {
+    it('formats hours only', () => {
+      const result = formatTimeFromSeconds(7200)
+      chai.expect(result).to.deep.equal('2 hours')
+    })
+
+    it('formats minutes only', () => {
+      const result = formatTimeFromSeconds(3540)
+      chai.expect(result).to.deep.equal('59 minutes')
+    })
+
+    it('formats seconds only', () => {
+      const result = formatTimeFromSeconds(59)
+      chai.expect(result).to.deep.equal('59 seconds')
+    })
+
+    it('formats hours and minutes', () => {
+      const result = formatTimeFromSeconds(7500)
+      chai.expect(result).to.deep.equal('2 hours, 5 minutes')
+    })
+
+    it('handles zero in seconds', () => {
+      const result = formatTimeFromSeconds(0)
+      chai.expect(result).to.deep.equal('0 seconds')
+    })
+
+    it('rounds seconds down to nearest minute if number is more than 60', () => {
+      const result = formatTimeFromSeconds(61)
+      chai.expect(result).to.deep.equal('1 minutes')
+    })
+
+    it('rounds seconds down to nearest minute if number is more than 3600 (an hour)', () => {
+      const result = formatTimeFromSeconds(3601)
+      chai.expect(result).to.deep.equal('1 hours')
     })
   })
 })
